move check_optim_state_in_sync to sanity check
